% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{Bayes}
\alias{Bayes}
\title{Bayes}
\usage{
Bayes(
  params,
  bunny,
  HPD = 0.95,
  P0 = TRUE,
  K = FALSE,
  probK = 0.8,
  PR = FALSE,
  R = FALSE,
  PS = FALSE,
  SaveTable = FALSE,
  plot = FALSE
)
}
\arguments{
\item{params}{A parameter file generated by the functions iCreateParam or CreateParam.}

\item{bunny}{A list generated by the function bunny containing all model specifications and posterior samples of inferences}

\item{HPD}{a scalar from 0 to 1 specifying the amount of probability within the highest posterior density interval to be calculated. This argument applies to Model mean, Treatments, Covariates, Contrasts and Variance Components.}

\item{P0}{Logical (default=TRUE). If TRUE, the probability of the estimate to be greater than 0 if the median is positive, or lower than 0 if the median is negative, will be computed. Note that in case of the contrasts between Treatments being ratios, prob >1 or prob<1 will be computed. This argument applies to all Treatments, Contrasts and Covariates.}

\item{K}{Logical (default=FALSE). If TRUE, a guaranteed value of the estimate with probability Kprob is computed. Only displayed if K has the same sign as the median of the posterior chain. This argument applies to all Treatments, Contrasts, Covariates and Variance Components.}

\item{probK}{a numeric value from 0 to 1 indicating the probability to consider in K. If K=TRUE and probK is not specified, a default value of 0.8 will be used.}

\item{PR}{Logical (default=FALSE). If TRUE, the probability of the posterior chain to be greater than a relevant value (R) if the median is positive, or lower than R if the median is negative, will be computed. This argument applies to Contrasts.}

\item{R}{a vector containing a relevant value per each trait. If PR=TRUE, this argument is mandatory. Note that in case of the contrasts between Treatments being ratios, R should be considered in Percentage (For example 1.1 for 10%). Suggested values are 1/3 of the sd of the trait for differences, and 10% for ratios.}

\item{PS}{Logical (default=FALSE). If TRUE, the probability of similarity of the estimate is computed. This is, the posterior probability that the estimate is between -R and R (or between 1/R and R if its a ratio)}

\item{SaveTable}{Logical (default=TRUE). If TRUE, inferences of the posterior distributions printed in the prompt are saved in an .csv file.}

\item{plot}{Logical (default=FALSE). If TRUE, posterior distributions of contrasts are plot and saved in a tiff file, highlighting P0 and PR if applicable.}
}
\value{
a list with the inferences computed for each trait
}
\description{
this function computes several features from posterior chains included in a list generated by the function bunny.
}
