% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bunny.R
\name{Bunny}
\alias{Bunny}
\title{Bunny}
\usage{
Bunny(params, Chain = FALSE)
}
\arguments{
\item{params}{A parameter file generated by the functions iCreateParam or CreateParam.}

\item{Chain}{Logical (default=FALSE). If TRUE, all the samples of the posterior distributions of the estimated parameters are stored for each trait.}
}
\value{
A list containing the samples of the posterior distributions of the estimated parameters.
}
\description{
this function fits a mixed-effects model with fixed effects, random effects, and interactions based on the provided parameter file. It conducts convergence diagnosis, computes LSMeans, and calculates contrasts between levels of treatments as indicated in the parameter file.
}
\examples{
# Example usage:
# bunny(params = my_param_file, Chain = TRUE)
# bunny(params = my_param_file, Chain = FALSE)

}
