% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeInferences.R
\name{ComputeInferences}
\alias{ComputeInferences}
\title{Title ComputeInferences}
\usage{
ComputeInferences(
  samples,
  HPD = 0.95,
  P0 = TRUE,
  K = FALSE,
  probK = 0.8,
  PR = FALSE,
  R = NULL,
  PS = FALSE,
  askCompare = "D"
)
}
\arguments{
\item{samples}{its an numeric vector containing samples from a posterior chain of any estimate}

\item{HPD}{a scalar from 0 to 1 specifying the amount of probability within the highest posterior density interval to be calculated.}

\item{P0}{Logical (default=TRUE). If TRUE, the probability of the estimate to be greater than 0 if the median is positive, or lower than 0 if the median is negative, will be computed.}

\item{K}{Logical (default=FALSE). If TRUE, a guaranteed value of the estimate with probability Kprob is computed. Only displayed if K has the same sign as the median of the posterior chain.}

\item{probK}{a numeric value from 0 to 1 indicating the probability to consider in K. If K=TRUE and probK is not specified, a default value of 0.8 will be used.}

\item{PR}{Logical (default=FALSE). If TRUE, the probability of the posterior chain to be greater than a relevant value (R) if the median is positive, or lower than R if the median is negative, will be computed.}

\item{R}{a scalar containing a relevant value for computing PR. If any PR or PS is TRUE, this argument is mandatory.}

\item{PS}{Logical (default=FALSE). If TRUE, the probability of similarity of the estimate is computed. This is, the posterior probability that the estimate is between -R and R.}

\item{askCompare}{can be either "D" or "R". If "D", PR is computed assuming -R as threshold if the median is <0, and PS is computed between -R and R. If "R", PR is computed assuming 1/R as threshold if the median is <1, and PS is computed between 1/R and R.}
}
\value{
a numeric vector with the Inferences of the posterior chains
}
\description{
this function makes inferences from posterior chains.
}
\examples{
 # Example usage:
 # Inferences <- computeInferences(samples)
 # Inferences <- computeInferences(samples, askCompare="R", P0=TRUE, K=TRUE, probK=0.8, PR=TRUE, R=1.1, PS=FALSE)
}
