% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateParam.R
\name{CreateParam}
\alias{CreateParam}
\title{CreateParam}
\usage{
CreateParam(
  file.name,
  na.codes = c("", "NA", "NULL"),
  hTrait = NULL,
  pTrait = NULL,
  hTreatment = NULL,
  pTreatment = NULL,
  askCompare = "D",
  hNoise = NULL,
  pNoise = NULL,
  hCov = NULL,
  pCov = NULL,
  hInter = NULL,
  pInter = NULL,
  typeInter = NULL,
  ShowInter = NULL,
  hRand = NULL,
  pRand = NULL,
  Seed = 1234,
  iter = 30000,
  burnin = 5000,
  lag = 10
)
}
\arguments{
\item{file.name}{is the file path of the data file (valid formats are .csv and .xlsx)}

\item{na.codes}{is a character of how missing values are coded}

\item{hTrait}{is a vector c() with the name of the traits separated by commas}

\item{pTrait}{is a vector c() with the column position of the traits separated by commas. User can specify traits using either hTrait, pTrait or both.}

\item{hTreatment}{is a vector c() with the name of the treatments effects separated by commas.}

\item{pTreatment}{is a vector c() with the column position of the treatments effects separated by commas. To specify treatment effects, user can enter either hTreatment, pTreatment or both.}

\item{askCompare}{The way you want to compare the levels of your Treatments. Is either "D" for difference, "R" for ratio, or "NA" if not applicable. Default value is D}

\item{hNoise}{is a vector c() with the name of the noise effects separated by commas.}

\item{pNoise}{is a vector c() with the column position of the noise effects separated by commas. User can specify noise effects using either hNoise, pNoise or both.}

\item{hCov}{is a vector c() with the name of the covariates}

\item{pCov}{is a vector c() with the column position of the covariates separated by commas. User can specify covariates using either hCov, pCov or both.}

\item{hInter}{is a matrix of dimensions n x 2 with n being the number of order 2 interactions to be considered, and the two columns indicate the name of the two interaction components (treatments, noise or covariates). For example, If considering two interactions between effects AE and Sex and AE and pH, argument needs to be entered as hInter=c("AE","Sex","AE","pH").}

\item{pInter}{is a matrix of dimensions n x 2 with n being the number of order 2 interactions to be considered, and the two columns indicate the column position of the two interaction components (treatments, noise or covariates). For example, If considering two interactions between effects positioned in columns 3 and 4 and 3 and 5, argument needs to be entered as pInter=c(3,4,3,5). User can specify noise effects using either hInter, pInter or both.}

\item{typeInter}{is a matrix of dimensions n x 2 with n being the number of order 2 interactions to be considered, and the two columns indicate if the components of the interaction are factors "F" or covariates "C".Its a mandantory argument if either hInter o pInter are not NULL. Example for two Interaction, one factor by factor and the other one factor by covariate, argument needs to be entered as typeInter=c("F","F","F","C").}

\item{ShowInter}{is a character vector specifying how interactions should be classified: either as treatments ("T") or as noise ("N"). Each element in the vector corresponds to an interaction defined in the model, and thus, the length of `ShowInter` must match the number of interactions specified. Its a mandantory argument if either hInter o pInter are not NULL.}

\item{hRand}{is a vector c() with the name of the random effects separated by commas.}

\item{pRand}{is a vector c() with the column position of the random effects separated by commas. User can specify random effects using either hRand, pRand or both.}

\item{Seed}{is a random number to initialize the sampling. Useful when aiming to get replicated results.}

\item{iter}{is the number of MCMC iterations. Default is 30 000.}

\item{burnin}{is the number of initial MCMC iterations to be discarded. Default is 5000.}

\item{lag}{is the thinning interval. Default is 10.}
}
\description{
Creates a parameter file from different arguments input by the user.
}
