% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataIMF.R
\docType{data}
\name{DataIMF}
\alias{DataIMF}
\title{DataIMF}
\format{
A data frame with 502 rows and 8 variables
}
\usage{
data(DataIMF)
}
\description{
The "DataIMF.csv" dataset (502 observations across 8 variables) comes from three generations of a divergent selection experiment for
intramuscular fat in rabbits.The data are based on perirenal fat (PFat), intramuscular fat (IMF) and loin pH of rabbits
belonging to both sexes (Sex), at constant live weight (LW). The data have been taken at
different seasons (AE, 2 levels) and rabbits were born at different parity orders (OP),
and have a random litter effect (c).
}
\keyword{datasets}
