% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBayes.R
\name{iBayes}
\alias{iBayes}
\title{iBayes}
\usage{
iBayes(params, bunny)
}
\arguments{
\item{params}{A parameter file generated by `iCreateParam` or `CreateParam`.}

\item{bunny}{A list produced by the `bunny` function, including all model specifications and posterior samples of inferences for different traits.}
}
\value{
Returns a list of computed inferences for each trait based on the user-defined criteria, including highest posterior density (HPD) intervals, probabilities of estimates being greater or different from specified values, and guaranteed values for a given probability.
}
\description{
This function facilitates interactive computation of posterior chain analysis with Bayesian Statistics. It is an interactive counterpart to `Bayes`, leveraging a list generated by the `Bunny` function. It guides the user through a series of questions to define analysis parameters dynamically.
}
\examples{
\dontrun{
# Example:
bayes_results <- iBayes(params=my_param_file, bunny=bunny_results)
}
}
