% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCreateParam.R
\name{iCreateParam}
\alias{iCreateParam}
\title{iCreateParam}
\usage{
iCreateParam()
}
\value{
The output from \code{\link{print}}
}
\description{
Creates a parameter file from interactive questions answered by the user.
}
\details{
Data file needs to be located in the working folder.
Data needs to be on .csv or .xlsx format. It admits missing values coded in different ways.
}
