% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabbitr.R
\name{rabbitr}
\alias{rabbitr}
\title{Establishing connection with RabbitMQ}
\usage{
rabbitr(host = "localhost", port = 5672, username = "guest",
  password = "guest")
}
\arguments{
\item{host}{Address of the RabbitMQ server.}

\item{port}{Port to connect to RabbitMQ server.}

\item{username}{Username for authentication.}

\item{password}{Password for authentication.}
}
\value{
Connection object
}
\description{
This is the main function of the package.
It establishes the connection between R and
the RabbitMQ server.
}
\examples{
\dontrun{
conn <- rabbitr(host='localhost', port=5672)
}

}
