% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeInferences.R
\name{ComputeInferences}
\alias{ComputeInferences}
\title{ComputeInferences}
\usage{
ComputeInferences(
  samples,
  HPD = 0.95,
  P0 = TRUE,
  K = FALSE,
  probK = 0.8,
  PR = FALSE,
  R = NULL,
  PS = FALSE,
  askCompare = "D"
)
}
\arguments{
\item{samples}{A numeric vector containing samples from a posterior distribution of any estimate.}

\item{HPD}{A scalar value between 0 and 1 specifying the desired probability to compute the highest posterior density (HPD) interval. Default value is 0.95.}

\item{P0}{A logical (default is `TRUE`). If `TRUE`, computes the probability that the posterior distribution is greater than zero if the median is positive, or less than zero if the median is negative.}

\item{K}{A logical(default is `FALSE`). If `TRUE`, computes a "guaranteed" value of the posterior distribution with a specified probability (`probK`). This value is only displayed if its sign matches the sign of the median of the posterior distribution.}

\item{probK}{A numeric value between 0 and 1 specifying the probability threshold for computing the guaranteed value when `K` is `TRUE`. If `K` is `TRUE` and `probK` is not specified, a default value of 0.80 is used.}

\item{PR}{A logical (default is `FALSE`). If `TRUE`, computes the probability that the posterior distribution is greater than a relevant value (`R`) if the median is positive, or less than `R` if the median is negative.}

\item{R}{A scalar specifying a relevant value for computing `PR`. It's mandatory if `PR` or `PS` is `TRUE`.}

\item{PS}{A logical (default is `FALSE`). If `TRUE`, computes the probability of similarity, i.e., the probability that the posterior distribution lies within `-R` to `R`}

\item{askCompare}{Specifies the mode of comparison, either "D" for differences or "R" for ratios, affecting how `PR` and `PS` are calculated. If "D", PR is computed assuming -R as threshold if the median is <0, and PS is computed between -R and R. If "R", PR is computed assuming 1/R as threshold if the median is <1, and PS is computed between 1/R and R.}
}
\value{
Returns a numeric vector containing calculated inferences from the posterior samples
}
\description{
This function calculates Statistical Inferences from Posterior Distributions
}
\examples{
\dontrun{
 # Example:
 # Basic inference calculation
 basic_inf <- computeInferences(samples)
 # Advanced inference calculation with custom settings
 advanced_inf <- computeInferences(samples, askCompare="R", HPD=0.7, P0=TRUE, K=TRUE,
 probK=0.8, PR=TRUE, R=1.1, PS=FALSE)
}
}
