% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBayes.R
\name{iBayes}
\alias{iBayes}
\title{iBayes}
\usage{
iBayes(params, bunny)
}
\arguments{
\item{params}{A parameter file generated by the functions iCreateParam or CreateParam.}

\item{bunny}{A list generated by the function bunny containing all model specifications and posterior samples of inferences}
}
\value{
a list with the inferences computed for each trait
}
\description{
this function computes several features from posterior chains included in a list generated by the function Bunny interactive questions answered by the user.
}
