% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rabbit.R
\name{Rabbit}
\alias{Rabbit}
\title{Title Rabbit}
\usage{
Rabbit()
}
\arguments{
\item{Chain}{Logical (default=FALSE). If TRUE, all the samples of the posterior distributions of the estimated parameters are stored for each trait.}
}
\value{
three lists: params (containing the parameter file), bunny (a list containing the samples of the posterior distributions of the estimated parameters for each trait), and inferences (a list with the inferences computed for each trait)
}
\description{
this function `runs general linear models and computes inferences from posterior chains.Its a didactic function as it has no arguments and all required information is interactively obtain through questions answered by the user
}
